# lilishop-master/docs_generated/im-api-openapi.json
{"openapi":"3.1.0","info":{"title":"im-api","version":"v3"},"servers":[{"url":"http://127.0.0.1:8885","description":"Generated server url"}],"paths":{"/store/seat/setting":{"get":{"tags":["åºéºç«¯,åå¸­è®¾ç½®"],"description":"æ¥è¯¢åå¸­è®¾ç½®","operationId":"getSetting","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageSeatSetting"}}}}}},"put":{"tags":["åºéºç«¯,åå¸­è®¾ç½®"],"description":"æ´æ°åå¸­è®¾ç½®","operationId":"update","parameters":[{"name":"seatSetting","in":"query","description":"åå¸­è®¾ç½®","required":true,"schema":{"$ref":"#/components/schemas/SeatSetting"}}],"responses":{"200":{"description":"OK"}}}},"/im/store/qa":{"put":{"tags":["ç®¡çç«¯,èªå®ä¹åè¯æ¥å£"],"description":"ä¿®æ¹èªå®ä¹é®ç­","operationId":"updateCustomWords","parameters":[{"name":"qa","in":"query","description":"é®ç­","required":true,"schema":{"$ref":"#/components/schemas/QA"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageQA"}}}}}},"post":{"tags":["ç®¡çç«¯,èªå®ä¹åè¯æ¥å£"],"description":"æ·»å é®ç­","operationId":"addCustomWords","parameters":[{"name":"qa","in":"query","description":"é®ç­","required":true,"schema":{"$ref":"#/components/schemas/QA"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageQA"}}}}}}},"/im/message/{id}":{"get":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"æ¥çImæ¶æ¯è¯¦æ","operationId":"get","parameters":[{"name":"id","in":"path","description":"Imæ¶æ¯ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageImMessage"}}}}}},"put":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"æ´æ°Imæ¶æ¯","operationId":"update_1","parameters":[{"name":"id","in":"path","description":"Imæ¶æ¯ID","required":true,"schema":{"type":"string"}},{"name":"imMessage","in":"query","description":"Imæ¶æ¯å¯¹è±¡","required":true,"schema":{"$ref":"#/components/schemas/ImMessage"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageImMessage"}}}}}}},"/im/message/clean/unred":{"put":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"æ¸é¤æææªè¯»æ¶æ¯","operationId":"cleanUnreadMessage","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageObject"}}}}}}},"/im/message":{"get":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"åé¡µè·åImæ¶æ¯","operationId":"historyMessage","parameters":[{"name":"messageQueryParams","in":"query","description":"æ¥è¯¢åæ°","required":true,"schema":{"$ref":"#/components/schemas/MessageQueryParams"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageListImMessage"}}}}}},"post":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"æ°å¢Imæ¶æ¯","operationId":"save","parameters":[{"name":"imMessage","in":"query","description":"Imæ¶æ¯å¯¹è±¡","required":true,"schema":{"$ref":"#/components/schemas/ImMessage"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageImMessage"}}}}}}},"/store/seat/setting/list":{"get":{"tags":["åºéºç«¯,åå¸­ç®¡ç"],"description":"åé¡µè·ååå¸­","operationId":"getSeats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageListSeatVO"}}}}}}},"/im/talk/{id}":{"get":{"tags":["èå¤©æ¥å£"],"summary":"æ¥çèå¤©è¯¦æ","operationId":"get_1","parameters":[{"name":"id","in":"path","description":"èå¤©ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageImTalk"}}}}}},"delete":{"tags":["èå¤©æ¥å£"],"summary":"å é¤èå¤©","operationId":"disable","parameters":[{"name":"id","in":"path","description":"èå¤©ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageObject"}}}}}}},"/im/talk/user/{uid}":{"get":{"tags":["èå¤©æ¥å£"],"summary":"æ¥çä¸æäººèå¤©è¯¦æ","operationId":"getUser","parameters":[{"name":"uid","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageImTalk"}}}}}}},"/im/talk/top":{"get":{"tags":["èå¤©æ¥å£"],"summary":"è®¾ç½®èå¤©ç½®é¡¶","operationId":"top","parameters":[{"name":"id","in":"query","description":"èå¤©ID","required":true,"schema":{"type":"string"}},{"name":"top","in":"query","description":"æ¯å¦ç½®é¡¶","required":true,"schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageObject"}}}}}}},"/im/talk/store/list":{"get":{"tags":["èå¤©æ¥å£"],"summary":"åé¡µè·ååå®¶èå¤©","operationId":"getStoreTalkList","parameters":[{"name":"imTalkQueryParams","in":"query","description":"èå¤©æ¥è¯¢åæ°","required":true,"schema":{"$ref":"#/components/schemas/IMTalkQueryParams"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageListImTalkVO"}}}}}}},"/im/talk/list":{"get":{"tags":["èå¤©æ¥å£"],"summary":"åé¡µè·åç¨æ·èå¤©","operationId":"getUserTalkList","parameters":[{"name":"imTalkQueryParams","in":"query","description":"èå¤©æ¥è¯¢åæ°","required":true,"schema":{"$ref":"#/components/schemas/IMTalkQueryParams"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageListImTalkVO"}}}}}}},"/im/talk/by/user/{userId}":{"get":{"tags":["èå¤©æ¥å£"],"summary":"æ¥çä¸æäººèå¤©è¯¦æ","operationId":"getByUser","parameters":[{"name":"userId","in":"path","description":"ç¨æ·ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageImTalkVO"}}}}}}},"/im/store/qa/page":{"get":{"tags":["ç®¡çç«¯,èªå®ä¹åè¯æ¥å£"],"description":"åé¡µè·åèªå®ä¹é®ç­","operationId":"getCustomWords","parameters":[{"name":"word","in":"query","description":"é®é¢","required":true,"schema":{"type":"string"}},{"name":"pageVo","in":"query","description":"åé¡µåæ°","required":true,"schema":{"$ref":"#/components/schemas/PageVO"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageIPageQA"}}}}}}},"/im/message/unreadMessage":{"get":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"è·åæææªè¯»æ¶æ¯","operationId":"getUnreadMessageCount","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageLong"}}}}}}},"/im/message/newMessage":{"get":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"æ¥çæ¯å¦ææ°æ¶æ¯","operationId":"hasNewMessage","parameters":[{"name":"accessToken","in":"query","description":"è®¿é®ä»¤ç","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageBoolean"}}}}}}},"/im/store/qa/{id}":{"delete":{"tags":["ç®¡çç«¯,èªå®ä¹åè¯æ¥å£"],"description":"å é¤èªå®ä¹é®ç­","operationId":"deleteCustomWords","parameters":[{"name":"id","in":"path","description":"é®ç­ID","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageString"}}}}}}},"/im/message/{ids}":{"delete":{"tags":["Imæ¶æ¯æ¥å£"],"summary":"å é¤Imæ¶æ¯","operationId":"delAllByIds","parameters":[{"name":"ids","in":"path","description":"Imæ¶æ¯IDåè¡¨","required":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultMessageObject"}}}}}}}},"components":{"schemas":{"SeatSetting":{"type":"object","description":"åå¸­è®¾ç½®","properties":{"tenantId":{"type":"string","description":"ç§æ·idid"},"welcome":{"type":"string","description":"æ¬¢è¿è¯­"},"outLineAutoReply":{"type":"string","description":"ç¦»çº¿èªå¨åå¤"},"longTermAutoReply":{"type":"string","description":"é¿æ¶é´èªå¨åå¤"}}},"QA":{"type":"object","description":"ç§æ·é®ç­","properties":{"tenantId":{"type":"integer","format":"int64","description":"ç§æ·id"},"question":{"type":"string","description":"é®é¢"},"answer":{"type":"string","description":"ç­æ¡"}}},"ResultMessageQA":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"$ref":"#/components/schemas/QA","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"ImMessage":{"type":"object","description":"Imæ¶æ¯","properties":{"fromUser":{"type":"string"},"toUser":{"type":"string"},"isRead":{"type":"boolean"},"messageType":{"type":"string","description":"è¯·ä¼ å­ç¬¦ä¸²æä¸¾å¼ï¼ä¸è¦ä¼  Swagger åè¡¨åé¢çæ°å­åºå·ã","enum":["MESSAGE","PICTURE","VOICE","GOODS","ORDER","VIDEO"]},"talkId":{"type":"string"},"text":{"type":"string"}}},"ResultMessageImMessage":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"$ref":"#/components/schemas/ImMessage","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"ResultMessageObject":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"ResultMessageSeatSetting":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"$ref":"#/components/schemas/SeatSetting","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"ResultMessageListSeatVO":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"type":"array","description":"çå®ä¸å¡æ°æ®è½½è·ã","items":{"$ref":"#/components/schemas/SeatVO"}}}},"SeatVO":{"type":"object","properties":{"tenantId":{"type":"string","description":"ç§æ·id"},"username":{"type":"string","description":"åå¸­ç¨æ·å"},"face":{"type":"string","description":"å®¢æ·å¤´å"},"password":{"type":"string","description":"åå¸­å¯ç "},"nickName":{"type":"string","description":"æµç§°"},"disabled":{"type":"boolean","description":"åå¸­ç¶æ"},"mobile":{"type":"string","description":"ææºå·ç ","minLength":1},"onlineStatus":{"type":"string"}},"required":["mobile"]},"ImTalk":{"type":"object","description":"èå¤©","properties":{"tenantId":{"type":"string","description":"åå¸­Id"},"userId1":{"type":"string","description":"ç¨æ·1 id"},"userId2":{"type":"string","description":"ç¨æ·2 id"},"top1":{"type":"boolean","description":"ç¨æ·1ç½®é¡¶"},"top2":{"type":"boolean","description":"ç¨æ·2ç½®é¡¶"},"disable1":{"type":"boolean","description":"ç¨æ·1 ä¸å¯è§"},"disable2":{"type":"boolean","description":"ç¨æ·2 ä¸å¯è§"},"name1":{"type":"string","description":"ç¨æ·1åå­"},"name2":{"type":"string","description":"ç¨æ·2åå­"},"face1":{"type":"string","description":"ç¨æ·1å¤´å"},"face2":{"type":"string","description":"ç¨æ·2å¤´å"},"storeFlag1":{"type":"boolean","description":"ç¨æ·1çåºéºæ è¯"},"storeFlag2":{"type":"boolean","description":"ç¨æ·2çåºéºæ è¯"},"lastTalkMessage":{"type":"string","description":"æåèå¤©åå®¹"},"lastMessageType":{"type":"string","description":"æååéæ¶æ¯ç±»å"},"tenantName":{"type":"string","description":"åå¸­åç§°"}}},"ResultMessageImTalk":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"$ref":"#/components/schemas/ImTalk","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"IMTalkQueryParams":{"type":"object","properties":{"userId1":{"type":"string","description":"ç¨æ·1 id"},"userId2":{"type":"string","description":"ç¨æ·2 id"},"name1":{"type":"string","description":"ç¨æ·1 name"},"name2":{"type":"string","description":"ç¨æ·2 name"},"userName":{"type":"string","description":"å³é®å­"}}},"ImTalkVO":{"type":"object","description":"èå¤©","properties":{"id":{"type":"string","description":"id"},"userId":{"type":"string","description":"ç¨æ· id"},"top":{"type":"boolean","description":"ç½®é¡¶"},"disable":{"type":"boolean","description":"ç¨æ· ä¸å¯è§"},"name":{"type":"string","description":"ç¨æ·åå­"},"face":{"type":"string","description":"ç¨æ·å¤´å"},"storeFlag":{"type":"boolean","description":"åºéºæ è¯"},"lastTalkMessage":{"type":"string","description":"æåèå¤©åå®¹"},"lastMessageType":{"type":"string","description":"æååéæ¶æ¯ç±»å"},"unread":{"type":"integer","format":"int64","description":"æªè¯»æ°é"}}},"ResultMessageListImTalkVO":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"type":"array","description":"çå®ä¸å¡æ°æ®è½½è·ã","items":{"$ref":"#/components/schemas/ImTalkVO"}}}},"ResultMessageImTalkVO":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"$ref":"#/components/schemas/ImTalkVO","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"PageVO":{"type":"object","description":"éç¨åé¡µæ¥è¯¢åæ°ã","properties":{"pageNumber":{"type":"integer","format":"int32","description":"é¡µç ï¼ä» 1 å¼å§ã","example":1},"pageSize":{"type":"integer","format":"int32","description":"æ¯é¡µæ¡æ°ã","example":10},"sort":{"type":"string","description":"æåºå­æ®µãé»è®¤æé©¼å³°å­æ®µèªå¨è½¬ä¸ºä¸åçº¿å­æ®µåã æåºå­æ®µãé»è®¤æé¡¹ç®çº¦å®è¿è¡é©¼å³°è½¬ä¸åçº¿ã","example":"createTime"},"order":{"type":"string","description":"æåºæ¹åï¼ä¼  asc æ descã","example":"desc"},"notConvert":{"type":"boolean","description":"æ¯å¦å³é­æåºå­æ®µçé©¼å³°è½¬ä¸åçº¿ãtrue è¡¨ç¤ºæåå­æ®µåç´æ¥æåºã","example":false}}},"IPageQA":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"current":{"type":"integer","format":"int64"},"total":{"type":"integer","format":"int64"},"records":{"type":"array","items":{"$ref":"#/components/schemas/QA"}},"pages":{"type":"integer","format":"int64","deprecated":true}}},"ResultMessageIPageQA":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"$ref":"#/components/schemas/IPageQA","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"MessageQueryParams":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32","description":"é¡µç ï¼ä» 1 å¼å§ã","example":1},"pageSize":{"type":"integer","format":"int32","description":"æ¯é¡µæ¡æ°ã","example":10},"sort":{"type":"string","description":"æåºå­æ®µãé»è®¤æé©¼å³°å­æ®µèªå¨è½¬ä¸ºä¸åçº¿å­æ®µåã","example":"createTime"},"order":{"type":"string","description":"æåºæ¹åï¼ä¼  asc æ descã","example":"desc"},"notConvert":{"type":"boolean","description":"æ¯å¦å³é­æåºå­æ®µçé©¼å³°è½¬ä¸åçº¿ãtrue è¡¨ç¤ºæåå­æ®µåç´æ¥æåºã","example":false},"talkId":{"type":"string"},"lastMessageId":{"type":"string"},"num":{"type":"integer","format":"int32"}}},"ResultMessageListImMessage":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"type":"array","description":"çå®ä¸å¡æ°æ®è½½è·ã","items":{"$ref":"#/components/schemas/ImMessage"}}}},"ResultMessageLong":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"type":"integer","format":"int64","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"ResultMessageBoolean":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"type":"boolean","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}},"ResultMessageString":{"type":"object","description":"ç»ä¸ååºåè£ãsuccess/code/message æè¿°ä¸å¡ç»æï¼result æ¿è½½å®éä¸å¡æ°æ®ã","properties":{"success":{"type":"boolean","description":"ä¸å¡æ¯å¦æåãtrue è¡¨ç¤ºæåï¼false è¡¨ç¤ºå¤±è´¥ã"},"message":{"type":"string","description":"ä¸å¡æç¤ºææ¡ãå¤±è´¥æ¶è¯·ä¼åç»å code å¤æ­ï¼ä¸è¦åªä¾èµ messageã"},"code":{"type":"integer","format":"int32","description":"ç»ä¸ä¸å¡ç¶æç ã200 è¡¨ç¤ºæåï¼å¶ä»å¼è¡¨ç¤ºå·ä½ä¸å¡å¤±è´¥åå ã","example":200},"timestamp":{"type":"integer","format":"int64","description":"æå¡ç«¯çæçååºæ¶é´æ³ï¼æ¯«ç§ã"},"result":{"type":"string","description":"çå®ä¸å¡æ°æ®è½½è·ã"}}}}}}